refactor(prior-authority): move PayloadFingerprint to domain-agnostic util package
Move PayloadFingerprint from command.application.priorauthority.data to
uk.gov.justice.laa.dstew.access.util as required by Step 11 of the plan.
Update PriorAuthorityDataStore and PriorAuthorityDataStoreTest imports accordingly.
