+++ data-access-service-axon/src/test/java/uk/gov/justice/laa/dstew/access/command/application/priorauthority/data/PriorAuthorityDataStoreTest.java
@@ -15,6 +15,7 @@ import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentCaptor;
 import org.springframework.dao.DataIntegrityViolationException;
 import uk.gov.justice.laa.dstew.access.content.priorauthority.PriorAuthorityContent;
+import uk.gov.justice.laa.dstew.access.util.PayloadFingerprint;
 
 class PriorAuthorityDataStoreTest {
